requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
transformers==4.36.0
pillow==10.1.0
pdf2image==1.16.3
python-docx==0.8.11
aiofiles==23.2.0
qwen-vl-utils==0.0.3
accelerate==0.25.0
flash-attn==2.3.3